+++ src/Data/Vec3/Tupled.hs
@@ -3,8 +3,8 @@
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TypeFamilies #-}
 
-module Data.Vec3.Unboxed.Tupled
-    ( TUVec3(..)
+module Data.Vec3.Tupled
+    ( TVec3(..)
     )
 
 where
@@ -16,7 +16,7 @@ import Data.Vector.Unboxed.Deriving
 import Data.Vec3.Class
 
 
--- | 'Vec3' implementation with "Data.Vector.Unboxed.Unbox" instance
+-- | 'Vec3' implementation with 'Data.Vector.Unboxed.Unbox' instance
 -- based on default Unbox instance for tuples of arrays, which wraps a
 -- vector of tuples as a tuple of vectors.
 --
@@ -28,37 +28,37 @@ import Data.Vec3.Class
 -- storage(z): [v1z-------+ ; v2z-------+  ...], length = N
 -- @
 --
--- You almost definitely want to use "Data.Vec3.Unboxed.UVec3" instead
--- as it has better performance.
-newtype TUVec3 = TUVec3 (Double, Double, Double)
+-- You almost definitely want to use 'CVec3' instead as it has better
+-- performance.
+newtype TVec3 = TVec3 (Double, Double, Double)
                deriving (Eq, Show)
 
 
-derivingUnbox "TUVec3"
-  [t|TUVec3 -> (Double, Double, Double)|]
-  [|\(TUVec3 v) -> v|]
-  [|TUVec3|]
+derivingUnbox "TVec3"
+  [t|TVec3 -> (Double, Double, Double)|]
+  [|\(TVec3 v) -> v|]
+  [|TVec3|]
 
 
-instance Vec3 TUVec3 where
-    newtype Matrix TUVec3 = UMatrix (TUVec3, TUVec3, TUVec3)
+instance Vec3 TVec3 where
+    newtype Matrix TVec3 = TMatrix (TVec3, TVec3, TVec3)
                            deriving (Eq, Show)
 
 
-    fromXYZ = TUVec3
+    fromXYZ = TVec3
     {-# INLINE fromXYZ #-}
 
-    toXYZ (TUVec3 v) = v
+    toXYZ (TVec3 v) = v
     {-# INLINE toXYZ #-}
 
-    fromRows (r1, r2, r3) = UMatrix (r1, r2, r3)
+    fromRows (r1, r2, r3) = TMatrix (r1, r2, r3)
     {-# INLINE fromRows #-}
 
-    toRows (UMatrix (r1, r2, r3)) = (r1, r2, r3)
+    toRows (TMatrix (r1, r2, r3)) = (r1, r2, r3)
     {-# INLINE toRows #-}
 
 
-derivingUnbox "UMatrix"
-  [t|Matrix TUVec3 -> (TUVec3, TUVec3, TUVec3)|]
-  [|\(UMatrix v) -> v|]
-  [|UMatrix|]
+derivingUnbox "TMatrix"
+  [t|Matrix TVec3 -> (TVec3, TVec3, TVec3)|]
+  [|\(TMatrix v) -> v|]
+  [|TMatrix|]